pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "" }]
classifiers = [
  "Programming Language :: Python :: 3",
]
dependencies = [
  "accelerate>=0.27.0",
  "datasets>=2.18.0",
  "openai>=1.40.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "torch>=2.1.0",
  "tqdm>=4.66.0",
  "transformers>=4.40.0",
  "trl>=0.9.0",
]

[project.urls]
Homepage = "https://example.com"

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["project"]
